% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPlot.R
\name{AccessPlot}
\alias{AccessPlot}
\title{Cheminement pour accéder aux placettes}
\usage{
AccessPlot(parc = NULL, CreateNetPP = T, Nb = 30, enrg = T, nom = "Acces.shp")
}
\arguments{
\item{parc}{= limites internes et externes (parcellaire, pistes, etc.).}

\item{CreateNetPP}{= Si TRUE (valeur par défaut) la fonction crée le réseau de placettes.
Si FALSE une boîte de dialogue permet de sélectionner le réseau de placettes.}

\item{Nb}{= nombre de placettes. Utile uniquement si CreateNetPP = TRUE.}

\item{enrg}{= Si TRUE (valeur par défaut) la fonction enregistre dans un fichier au format .shp
le point le plus proche de chaque placette situé sur une limite interne ou externe.}

\item{nom}{: nom du fichier (par défaut nom="Acces.shp"). Utile uniquement si enrg = TRUE.}
}
\value{
La fonction renvoie un tableau indiquant par placette la distance à parcourir.
Elle sauvegarde également le point de départ situé sur la desserte.
}
\description{
La fonction permet, à partir de la localisation des placettes et de la desserte
(routes, pistes, chemins), de déterminer le cheminement
}
\author{
Bruciamacchie Max
}
