% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateNetPP.R
\name{CreateNetPP}
\alias{CreateNetPP}
\title{Création d'un réseau de placettes permanentes}
\usage{
CreateNetPP(
  perim = NULL,
  Nb = 30,
  Cellsize = NULL,
  Type = "regular",
  enrg = T,
  nom = "MailleAut.shp"
)
}
\arguments{
\item{perim}{= objet au format sf. Si NULL ouverture d'une boîte de dialogue permettant de sélectionner
un fichier shp contenant des POLYGON.}

\item{Nb}{= nombre souhaité de placettes}

\item{Cellsize}{= pas de la maille en mètre}

\item{Type}{= type de maille souhaitée. Les options possibles sont : random, regular
(option par défaut), stratified, nonaligned, hexagonal}

\item{enrg}{= Si TRUE (valeur par défaut) la fonction enregistre la grille de points.}

\item{nom}{= nom souhaité pour le fichier ("MailleAut" par défaut)}
}
\value{
La fonction construit le fichier géoréférencé des placettes dans le même dossier
que le ficheir en entrée avec "MailleAut" comme nom par défaut.
}
\description{
La fonction construit un réseau de placettes permanentes en fonction
du périmètre de la zone d'étude, du nombre de placettes souhaité et de la forme du réseau.
}
\author{
Bruciamacchie Max
}
