% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePerim.R
\name{CreatePerim}
\alias{CreatePerim}
\title{Création d'un périmètre autour des points de placettes}
\usage{
CreatePerim(points = NULL, TailleBuffer = 100, nom = "PerimAut.shp")
}
\arguments{
\item{points}{= objet sf ou fichier POINT au format .shp}

\item{TailleBuffer}{= Taille du buffer en mètre (50m par défaut)}

\item{nom}{= nom souhaité pour le fichier en sortie.}
}
\value{
La fonction sauvegarde un SpatialPolygonDataFrame.
}
\description{
La fonction utilise en entrée un shape de type ponctuel contenant les placettes permanentes
d’une ou plusieurs forêts. Le fichier en entrée est un SpatialPointsDataFrame dont la table attributaire
contient au moins 2 champs. Le premier permet de distinguer les dispositifs, le second les placettes.
Les périmètres sont construits à l'aide de la fonction gConvexHull du package rgeos.
}
\author{
Bruciamacchie Max
}
