% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSV2SHP.R
\name{CSV2SHP}
\alias{CSV2SHP}
\title{Conversion CSV to SHP}
\usage{
CSV2SHP(epsg = 2154, nom = "Placettes.shp")
}
\arguments{
\item{epsg}{= code EPSG (système de projection)}

\item{nom}{= nom du fichier en sortie. Par défaut "Placettes.shp"}
}
\value{
La fonction renvoie un objet au format sf mais l'enregistre dans le format
indiqué dans l'extension de l'argument nom.
Le fichier est enregistré dans le répertoire en cours.
}
\description{
La fonction permet de convertir un fichier de points au format CSV
en fichier géoréférencé au format SHP
}
\author{
Bruciamacchie Max
}
