% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvIFNmult.R
\name{CvIFNmult}
\alias{CvIFNmult}
\title{CvIFNmult}
\usage{
CvIFNmult(perim = NULL, rep = NULL, TailleBuffer = NULL, enreg = F)
}
\arguments{
\item{perim}{= objet au format sf. Si NULL ouverture d'une boîte de dialogue permettant de sélectionner
un fichier shp contenant des POLYGON.}

\item{TailleBuffer}{= Taille du buffer en mètre. Elle est calculée de manière à ce que la zone de
recherche des placettes IFN soit au moins égale à 2000 ha. Elle est au minimum égale à 450 m.}

\item{enreg}{= argument permettant de choisir l'enregistrement ou pas du tableau au format .xlsx.
Par défaut enreg=FALSE.}
}
\value{
La fonction renvoie un tableau par groupe des coefficients de variation
sur la surface terrière
\itemize{
 \item des essences contribuant à plus de 10\% de la surface terrière
 \item des PER, PB, BM et GB
 \item des gros bois de l'essence contribuant le plus à la surface terrière
}
}
\description{
La fonction renvoie des coefficients de variation sur la surface terrière
en utilisant la base de données arbres de l'IFN. Par rapport à la fonction CvIFN,
elle permet de raisonner par secteur, groupe de parcelle, regroupement de polygone, etc.
La fonction nécessite en entrée un shape correspondant au périmètre retenu (forêts, massifs,
etc.). Une boîte de dialogue permet de choisir le champ de regroupement.
}
\details{
Coefficients de variation pour des groupes
}
\examples{
library(tidyverse)
library(sf)
library(DataForet)
library(PPtools)
#
res <- CvIFNCvIFNmult()
res$tab
res$Nbres

}
\author{
Bruciamacchie Max
}
