% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions.R
\name{ExtractPlac}
\alias{ExtractPlac}
\title{Extraction des placettes IFN au voisinage foret}
\usage{
ExtractPlac(shp, TailleBuffer = NULL)
}
\arguments{
\item{shp}{= périmètre sous forme de polygone.}

\item{TailleBuffer=}{taille du buffer. par défaut TailleBuffer=NULL}
}
\value{
La fonction renvoie un polygone.
}
\description{
Extraction des placettes IFN au voisinage foret
}
\author{
Bruciamacchie Max
}
