% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions.R
\name{Find_Verif_point}
\alias{Find_Verif_point}
\title{##############################
Recherche et vérification de POINT}
\usage{
Find_Verif_point(points = NULL)
}
\arguments{
\item{poly}{= périmètre sous forme de polygone.}
}
\value{
La fonction renvoie un polygone.
}
\description{
##############################
Recherche et vérification de POINT
}
\author{
Bruciamacchie Max
}
