% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fonctions.R
\name{Find_Verif_poly}
\alias{Find_Verif_poly}
\title{##############################
Recherche et vérification de POLYGON}
\usage{
Find_Verif_poly(poly = NULL)
}
\arguments{
\item{poly}{= périmètre sous forme de polygone.}
}
\value{
La fonction renvoie un objet POLYGON ou MULTIPOLYGON
}
\description{
Choix du périmètre et vérification que le fichier retenu est constitué de
polygone
}
\author{
Bruciamacchie Max
}
