% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFNacctD.R
\name{IFNacctD}
\alias{IFNacctD}
\title{Accroissement sur le diamètre}
\usage{
IFNacctD(perim = NULL, SeuilCircf = 50, SeuilNb = 10, UseSer = T, enreg = F)
}
\arguments{
\item{perim}{= objet au format sf. Si NULL ouverture d'une boîte de dialogue permettant de sélectionner
un fichier shp contenant des POLYGON.}

\item{SeuilCircf}{= seuil minimal de circonférence en cm en dessous duquel l'arbre n'est pas retenu
pour le calcul du tarif. Par défaut ce seuil est fixé à 50 cm.}

\item{SeuilNb}{= seuil minimal du nombre d'arbres dans la base IFN pour qu'une essence soit retenue.
Par défaut ce seuil est fixé à 10 tiges.}

\item{UseSer}{= argument permettant de choisir si le calcul des numéros de tarif doit se faire
au sein d'une même sylvoécorégion. Par défaut, UseSer=TRUE.}

\item{enreg}{= argument permettant de choisir l'enregistrement ou pas du tableau au format .csv.
Par défaut enreg=FALSE.}
}
\value{
La fonction renvoie un tableau des accroissements sur le diamètre par essence et classe de diamètre.
Elle intègre les coefficients d'écorce, les accroissements sont donc sur écorce.
}
\description{
La fonction renvoie les accroissements sur le diamètre par essence et classe de diamètre
en utilisant la base de données arbres de l'IFN.
La fonction nécessite en entrée un shape correspondant au périmètre retenu : forêt, massif,
sylvoécorégion, etc.
}
\examples{
data("FD")
perim = FD \%>\% filter(IIDTN_FRT == "F10451Y")
acct <- IFNacctD(perim)
# résultats
acct$tab
acct$Effectif
acct$Graph

}
\author{
Bruciamacchie Max
}
