% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmSmallPolys.R
\name{RmSmallPolys}
\alias{RmSmallPolys}
\title{Simplification du domaine d'étude - suppression des trous au sein des polygones}
\usage{
RmSmallPolys(perim = NULL, nom = "PerimDef.shp")
}
\arguments{
\item{perim}{= objet au format sf. Si NULL ouverture d'une boîte de dialogue permettant de sélectionner
un fichier shp contenant des POLYGON.}

\item{nom}{= nom souhaité pour le fichier en sortie ("PerimDef" par défaut).}

\item{minarea}{= seuil minimal de surface (2500 m2 par défaut)}
}
\value{
La fonction renvoit un fichier au format .shp.
}
\description{
La fonction RmSmallPolys() utilise en entrée un shape de type surfacique.\cr
Elle permet de supprimer des polygones de surface inférieure à un seuil choisi par l'opérateur.\cr
Elle n'a de sens qu'après l'utilisation de la fonction UnionPolyDeleteHole()
}
\author{
Bruciamacchie Max
}
