% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarifChaude.R
\name{TarifChaude}
\alias{TarifChaude}
\title{Tarifs de cubage Chaudé}
\usage{
TarifChaude(num = 13, diam = 45, haut = 13)
}
\arguments{
\item{num}{= numéro de tarif. Ce numéro permet d'établir le lien avec la décroissance par classe de diamètre.}

\item{diam}{= diamètre(s) à 1,3 m (en cm).}

\item{haut}{= hauteur(s) découpe (en m);}
}
\value{
La fonction renvoie le volume calculé avec un tarif Chaudé à deux entrées. Le résultat est exprimé en m3.
}
\description{
Calcul du volume d'un arbre ou d'une grume à l'aide d'un tarif Chaudé.
}
\examples{
\donttest{
Hauts <- c(13,14,14)
Diams = c(45,45,50)
TarifChaude(13, Diams, Hauts)
}
}
\author{
Bruciamacchie Max
}
