% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarifErreurSch.R
\name{TarifErreurSch}
\alias{TarifErreurSch}
\title{Estimation de l'erreur dans l'utilisation d'un tarif Schaeffer}
\usage{
TarifErreurSch(df1, df2)
}
\arguments{
\item{df1}{= data.frame correspondant à l'échantillon d'arbres sur lequel les numéros de tarif Schaeffer
seront recherchés. Ce tableau doit contenir au moins 2 colonnes nommées Diam et Volume.}

\item{df2}{= data.frame correspondant au lot d'arbres à cuber. Ce second tableau doit contenir au moins
2 colonnes nommées Diam et Nombre.}
}
\value{
La fonction renvoie un tableau qui fournit par type de tarif Schaeffer rapide, intermédiaire,
lent ou très lent le numéro à utiliser, le volume du lot accompagné de son erreur d'estimation.
}
\description{
La fonction calcule l'erreur liée à l'estimation du volume d'un lot par utilisation d'un tarif
Schaeffer
}
\examples{
\donttest{
library(dplyr)
data(Echan)
data(Lot)
TarifErreurSch(Echan,Lot)
}

}
\author{
Bruciamacchie Max
}
