% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarifIFN.R
\name{TarifIFN}
\alias{TarifIFN}
\title{Tarif de cubage volume géométrique bois fort}
\usage{
TarifIFN(
  perim = NULL,
  SeuilCircf = 50,
  SeuilNb = 10,
  pas = 1000,
  UseSer = F,
  enreg = F
)
}
\arguments{
\item{perim}{= objet au format sf. Si NULL ouverture d'une boîte de dialogue permettant de sélectionner
un fichier shp contenant des POLYGON. Par défaut perim=NULL.}

\item{SeuilCircf}{= seuil minimal de circonférence en cm en dessous duquel l'arbre n'est pas retenu
pour le calcul du tarif. Par défaut le seuil est fixé à 50 cm.}

\item{SeuilNb}{= seuil minimal du nombre d'arbres dans la base IFN pour qu'une essence soit retenue.
Par défaut ce seuil est égal à 10.}

\item{pas}{= résolution (par défaut = 1000 m)}

\item{UseSer}{= argument permettant de choisir si le calcul des numéros de tarif doit se faire
au sein d'une même sylvoécorégion. Par défaut, UseSer=TRUE.}

\item{enreg}{= argument permettant de choisir l'enregistrement ou pas du tableau au format .xlsx.
Par défaut enreg=FALSE.}
}
\value{
La fonction renvoie un tableau des types et numéros de tarif Schaeffer par essence.
}
\description{
La fonction renvoie un tableau des types et numéros de tarif Schaeffer par essence
en utilisant la base de données arbres de l'IFN. Il s'agit d'un volume géométrique bois fort tige.
La fonction nécessite en entrée un shape correspondant au périmètre retenu : forêt, massif,
sylvoécorégion, etc.
}
\author{
Bruciamacchie Max
}
