% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarifSch.R
\name{TarifSch}
\alias{TarifSch}
\title{Tarifs de cubage Schaeffer}
\usage{
TarifSch(type = "SchR", num = 9, diam = 45)
}
\arguments{
\item{type}{= type de tarif. Par défaut type = "SchR", sinon "SchI", "SchL" ou "SchTL"}

\item{num}{= numéro de tarif.}

\item{diam}{= diamètre à 1,3 m (en cm).}
}
\value{
La fonction renvoie le volume calculé avec un tarif Schaeffer à une entrée, que ce soit un tarif
rapide, intermédiaire, lent ou très lent. Le résultat est exprimé en m3.
}
\description{
Calcul du volume d'un arbre ou d'une grume à l'aide d'un tarif Schaeffer à une entrée.
}
\examples{
\donttest{
# -------- Tarifs Schaeffer
Types = c("SchR","SchR","SchL")
Types2 = c("SchR","Sch","SchL")
Nums = c(9,8,9)
Diams = c(45,45,50)
TarifSch(Types,Nums,Diams)
TarifSch(Types2,Nums,Diams)
}

}
\author{
Bruciamacchie Max
}
