% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TarifSch2.R
\name{TarifSch2}
\alias{TarifSch2}
\title{Tarifs de cubage Schaeffer2}
\usage{
TarifSch2(num = 2, diam = 45, haut = 13)
}
\arguments{
\item{num}{= numéro de tarif. Ce numéro doit être l'un des suivants :
c(1,2,3,4.1,4.2,5.1,5.2,6,7,8.1,8.2,9,10).}

\item{diam}{= diamètre à 1,3 m (en cm).}

\item{haut}{= hauteur découpe (en m);}
}
\value{
La fonction renvoie le volume calculé avec un tarif Schaeffer à deux entrées. Le résultat est exprimé en m3.
}
\description{
Calcul du volume d'un arbre ou d'une grume à l'aide d'un tarif Schaeffer à 2 entrées.
}
\examples{
\donttest{
Hauts <- c(13,14,14)
Diams = c(45,45,50)
TarifSch2(2, Diams, Hauts)
}
}
\author{
Bruciamacchie Max
}
