% Generated by roxygen2 (4.0.2): do not edit by hand
\name{UnionPolyDeleteHole}
\alias{UnionPolyDeleteHole}
\title{Union des différents polygones d'un shape et suppression des trous.}
\usage{
UnionPolyDeleteHole()
}
\value{
La fonction construit le shape PerimCor.shp dans le même dossier.
}
\description{
La fonction utilise en entrée un shape de type surfacique.
Elle aggrège tous les polygones en un seul, et supprime les éventuels trous.
}
\author{
Bruciamacchie Max
}

