% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WritePointGPX.R
\name{WritePointGPX}
\alias{WritePointGPX}
\title{Ecriture au format GPX de données pontuelles}
\usage{
WritePointGPX(points = NULL, nom = "gpxfile.gpx")
}
\arguments{
\item{points}{= objet sf ou fichier POINT au format .shp}

\item{nom}{= nom du fichier en sortie. Par défaut nom="gpxfile.gpx".}
}
\value{
La fonction enregistre dans le même répertoire que le fichier utlisé en entrée.
}
\description{
La fonction sauvegarde au format GPX des données ponctuelles en permettant de choisir
dans la table attributaire, par le biais d'une boîte de dialogue, le champ qui sera utilisé
pour nommer les points.
}
\examples{
WritePointGPX()

}
\author{
Bruciamacchie Max
}
