% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CreateNetPP}
\alias{CreateNetPP}
\title{Création d'un réseau de placettes permanentes}
\usage{
CreateNetPP(Nb, Type = "regular", Objet = F)
}
\arguments{
\item{Nb}{= nombre souhaité de placettes}

\item{Type}{= type de maille souhaitée. Les options possibles sont : random, regular
(option par défaut), stratified, nonaligned, hexagonal}

\item{Objet}{= FALSE par défaut. Si True la fonction retourne la grille de points}
}
\value{
La fonction construit le shape Maille.shp dans un dossier choisi par l'utilisateur.
}
\description{
La fonction construit un réseau de placettes permanentes en fonction
du périmètre de la zone d'étude, du nombre de placettes souhaité et de la forme du réseau
}
\author{
Bruciamacchie Max
}

