% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CreatePerim}
\alias{CreatePerim}
\title{Création d'un périmètre autour des points de placettes}
\usage{
CreatePerim(TailleBuffer = 50, enreg = T)
}
\arguments{
\item{TailleBuffer}{= Taille du buffer en mètre (50m par défaut)}

\item{enreg}{= argument permettant de choisir l'enregistrement ou pas du fichier au format .shp.
Par défaut enreg=TRUE.}
}
\value{
La fonction renvoie un tableau des types et numéros de tarif Schaeffer par essence.
}
\description{
La fonction un SpatialPolygoneDataFrame contenant les périmètres des forêts.
Ils sont construits à l'aide de la fonction gConvexHull du package. Le fichier en entrée est un
SpatialPointsDataFrame dont la table attributaire contient au moins 2 champs. Le premier permet de distinguer
les dispositifs, le second les placettes
}
\author{
Bruciamacchie Max
}

