% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TarifIFN}
\alias{TarifIFN}
\title{Tarif de cubage volume géométrique bois fort}
\usage{
TarifIFN(TailleBuffer = 30000, SeuilCircf = 50, SeuilNb = 10,
  res = 1000, UseSer = T, enreg = F)
}
\arguments{
\item{TailleBuffer}{= Taille du buffer en mètre.}

\item{SeuilCircf}{= seuil minimal de circonférence en cm en dessous duquel l'arbre n'est pas retenu
pour le calcul du tarif.}

\item{SeuilNb}{= seuil minimal du nombre d'arbres dans la base IFN pour qu'une essence soit retenue.}

\item{res}{= résolution (par défaut = 1000 m)}

\item{UseSer}{= argument permettant de choisir si le calcul des numéros de tarif doit se faire
au sein d'une même sylvoécorégion. Par défaut, UseSer=TRUE.}

\item{enreg}{= argument permettant de choisir l'enregistrement ou pas du tableau au format .xlsx.
Par défaut enreg=False.}
}
\value{
La fonction renvoie un tableau des types et numéros de tarif Schaeffer par essence.
}
\description{
La fonction renvoie un tableau des types et numéros de tarif Schaeffer par essence
en utilisant la base de donées arbres de l'IFN. Il s'agit d'un volume géométrique bois fort tige.
La fonction nécessite en entrée un shape correspondant au périmètre retenu : forêt, massif,
sylvoécorégion, etc.
}
\author{
Bruciamacchie Max
}

